+++ src/lib_mockup/local_services.ml
@@ -952,10 +952,10 @@ module Make (E : MENV) = struct
     @@ Directory.register
          Directory.empty
          E.Block_services.S.Operations.operations
-         (fun (((), chain), _block) _query () ->
+         (fun (((), chain), _block) query () ->
            with_chain ~caller_name:"operations" chain (fun () ->
                (* FIXME: Better answer here *)
-               Tezos_rpc.Answer.return [[]; []; []; []]))
+               Tezos_rpc.Answer.return (query#version, [[]; []; []; []])))
 
   let monitor_operations () =
     let open Lwt_syntax in